Refactor: libpacemaker: move location functions to own source file
The location functions from pcmk_sched_constraints.c (which was still >2K
lines) are now in a new file pcmk_sched_location.c.

There are no code changes aside from function renames, formatting, and
comments.
