+++ lib/pacemaker/libpacemaker_private.h
@@ -50,6 +50,9 @@ gboolean pcmk__tag_to_set(xmlNode *xml_obj, xmlNode **rsc_set, const char *attr,
                           gboolean convert_rsc, pe_working_set_t *data_set);
 
 G_GNUC_INTERNAL
+void pcmk__unpack_location(xmlNode *xml_obj, pe_working_set_t *data_set);
+
+G_GNUC_INTERNAL
 void pcmk__unpack_colocation(xmlNode *xml_obj, pe_working_set_t *data_set);
 
 G_GNUC_INTERNAL
